style: add typescript and fix lint errors
