--- sample/scripts/react-prebuild.js
@@ -1,65 +1,0 @@
-const path = require('path');
-const { exec } = require('child_process');
-const fs = require('fs');
-const rimraf = require('rimraf');
-const { type } = require('os');
-
-function renameOutputFolder(buildFolderPath, outputFolderPath) {
-    return new Promise((resolve, reject) => {
-        fs.rename(buildFolderPath, outputFolderPath, (err) => {
-            if (err) {
-                reject(err);
-            } else {
-                resolve('Successfully built!');
-            }
-        });
-    });
-}
-
-function execPostReactBuild(buildFolderPath, outputFolderPath) {
-    return new Promise((resolve, reject) => {
-        if (fs.existsSync(buildFolderPath)) {
-            if (fs.existsSync(outputFolderPath)) {
-                rimraf.rimraf(outputFolderPath, []).then(() => {
-                    renameOutputFolder(buildFolderPath, outputFolderPath)
-                      .then(val => resolve(val))
-                      .catch(e => reject(e));
-                }).catch((err) => {
-                  if (err) {
-                      reject(err);
-                      return;
-                  }
-              });
-            } else {
-                renameOutputFolder(buildFolderPath, outputFolderPath)
-                    .then(val => resolve(val))
-                    .catch(e => reject(e));
-            }
-        } else {
-            reject(new Error('build folder does not exist'));
-        }
-    });
-}
-
-module.exports = () => {
-    const projectPath = path.resolve(process.cwd(), './node_modules/.bin/react-scripts');
-    return new Promise((resolve, reject) => {
-        exec(`${projectPath} build`,
-            (error) => {
-                if (error) {
-                    console.error(error);
-                    reject(error);
-                    return;
-                }
-                execPostReactBuild(path.resolve(__dirname, '../build/'), path.join(__dirname, '../www/'))
-                    .then((s) => {
-                        console.log(s);
-                        resolve(s);
-                    })
-                    .catch((e) => {
-                        console.error(e);
-                        reject(e);
-                    });
-            });
-    });
-};
