+++ sample/src/App.tsx
@@ -1,14 +1,12 @@
-import logo from './logo.svg';
+import * as React from 'react';
 import './App.css';
 
-function App() {
-  console.log(`TESTEST App`);
+export default function App() {
   return (
     <div className="App">
       <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
         <p>
-          Edit <code>src/App.js</code> and save to reload.
+          Hello.
         </p>
         <a
           className="App-link"
@@ -22,5 +20,3 @@ function App() {
     </div>
   );
 }
-
-export default App;
