--- sample/src/index.js
@@ -1,28 +1,0 @@
-import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './index.css';
-import App from './App';
-import reportWebVitals from './reportWebVitals';
-
-const renderReactDom = () => {
-  console.log(`TESTEST renderReactDom`);
-  const root = ReactDOM.createRoot(document.getElementById('root'));
-  root.render(
-    <React.StrictMode>
-      <App />
-    </React.StrictMode>
-  );
-}
-
-if (window.cordova) {
-  document.addEventListener('deviceready', () => {
-    renderReactDom();
-  }, false);
-} else {
-  renderReactDom();
-}
-
-// If you want to start measuring performance in your app, pass a function
-// to log results (for example: reportWebVitals(console.log))
-// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
-reportWebVitals();
